requirements.txt
streamlit
pandas
joblib
rdkit-pypi
scikit-learn
matplotlib
seaborn
